+++ src/App.tsx
@@ -1,6 +1,6 @@
 // Redux
 import { Provider } from "react-redux";
-import { PersistGate } from 'redux-persist/integration/react'
+import { PersistGate } from "redux-persist/integration/react";
 import { store, persistor } from "./providers/redux";
 
 // Containers