+++ src/components/PostContent/PostContent.test.tsx
@@ -2,14 +2,16 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 
 import { PostContent } from ".";
+import { IPost } from "../../dal/types";
 
-const post = {
+const post: IPost = {
   id: "1",
   title: "Post 1",
   author: "mserpez",
   createdAt: new Date().getTime(),
   commentsQty: 1000000,
   photoURL: "",
+  photoURLBig: "",
   thumbnail: "",
 };