+++ src/components/PostImage/PostImage.test.tsx
@@ -3,10 +3,10 @@ import "@testing-library/jest-dom/extend-expect";
 
 import { PostImage } from ".";
 
-test("Renders the component correctly", () => {
+test("Renders the component correctly", async () => {
   // Render component
   render(<PostImage />);
 
   // Find the element
-  screen.getByTestId("post-image");
+  await screen.findByRole("img");
 });