+++ src/components/PostImage/index.tsx
@@ -1,8 +1,6 @@
 import Image from "react-graceful-image";
 import styled from "@emotion/styled";
 
-import loadingImg from "../../assets/loading.gif";
-
 export enum PostImageSize {
   "sm" = "110",
   "md" = "400",
@@ -12,8 +10,6 @@ export enum PostImageSize {
 export interface PostImageProps extends PostImageBaseProps {
   url?: string;
   alt?: string;
-  onImageClick?: () => void;
-  loading?: boolean;
 }
 
 interface PostImageBaseProps {
@@ -23,15 +19,14 @@ interface PostImageBaseProps {
 export function PostImage({
   url,
   alt,
-  onImageClick,
   size = PostImageSize.sm,
-  loading = false,
 }: PostImageProps) {
-  if (loading) return <StyledLoadingImage alt="Loading..." src={loadingImg} />;
-
   return (
-    <div data-testid="post-image-root" onClick={onImageClick}>
-      <StyledImage src={url} width={size} alt={alt} />
+    <div data-testid="post-image-root">
+      {/**
+       *  noLazyLoad: TODO: Check this prop/component, lazy load is not working fine sometimes.
+       */}
+      <StyledImage src={url} width={size} alt={alt} noLazyLoad />
     </div>
   );
 }
@@ -40,10 +35,8 @@ const StyledImage = styled(Image)`
   max-height: 100%;
   margin-right: 12px;
   margin-top: 12px;
-`;
-
-const StyledLoadingImage = styled.img`
-  width: 200px;
-  height: 200px;
-  margin: 24px;
+  @media (max-width: 776px) {
+    width: 90%;
+    margin-left: 5%;
+  }
 `;