+++ src/components/PostItem/PostItem.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from "@storybook/react/types-6-0";
 
-import { PostItem as PostItemComponent, PostItemProps } from ".";
+import { PostItem as PostItemComponent, IPostItemProps } from ".";
 
 export default {
   title: "Components/Post/Item",
@@ -12,7 +12,7 @@ export default {
   },
 } as Meta;
 
-const Template: Story<PostItemProps> = (args) => (
+const Template: Story<IPostItemProps> = (args) => (
   <PostItemComponent {...args} />
 );
 
@@ -25,6 +25,7 @@ PostItem.args = {
     createdAt: new Date().getTime(),
     commentsQty: 1000000,
     photoURL: "",
+    photoURLBig: "",
     thumbnail: "",
   },
 };