+++ src/containers/Reddit/Container.tsx
@@ -5,13 +5,20 @@ import { useDispatch } from "react-redux";
 import {
   fetchPostsAction,
   readPostAction,
+  toggleMenuAction,
   dismissPostAction,
   dismissAllPostsAction,
 } from "./actions";
 // Internal container components
 import Content from "./Content";
 // Reddit Selectors
-import { useOpenedPost, usePostList, useIsServiceLoading, useServiceError } from "./selectors";
+import {
+  useOpenedPost,
+  usePostList,
+  useIsServiceLoading,
+  useServiceError,
+  useIsOpenMenu,
+} from "./selectors";
 
 export function Reddit() {
   const dispatch = useDispatch();
@@ -19,9 +26,14 @@ export function Reddit() {
   const openedPost = useOpenedPost();
   const isServiceLoading = useIsServiceLoading();
   const serviceError = useServiceError();
+  // It will apply only on mobile resolution.
+  const isMenuOpen = useIsOpenMenu();
 
   // Is in charge to mark a post as read, to hide Badge.
   const readPost = (id: string) => dispatch(readPostAction(id));
+
+  // Once user click on post, then close menu to visualize it;
+  const toggleMenu = (isOpen: boolean) => dispatch(toggleMenuAction(isOpen));
 
   // Is in charge to set dismissed flag, to filter and not show it.
   const dismissPost = (id: string) => dispatch(dismissPostAction(id));
@@ -46,6 +58,8 @@ export function Reddit() {
       dismissAllPosts={dismissAllPosts}
       isServiceLoading={isServiceLoading}
       serviceError={serviceError}
+      isMenuOpen={isMenuOpen}
+      toggleMenu={toggleMenu}
     />
   );
 }