+++ src/dal/types.ts
@@ -4,6 +4,7 @@ export interface IPost {
   createdAt: number;
   id: string;
   photoURL: string;
+  photoURLBig: string;
   thumbnail: string;
   title: string;
   // custom attrs